vector-projection-3d/backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
psycopg-binary==3.2.9
psycopg==3.2.9
numpy==2.2.6
pandas==2.3.1
umap-learn==0.5.9.post2
scikit-learn==1.7.1
networkx==3.5
